Move docs list to separate module We want to pull RN API docs into exponent. So moving them into a module will be useful since we don't want to keep a duplicate list.
